Extract draft pair optimizer
